BREAKING CHANGE last changes in presentation of the library
