+++ README.md
@@ -47,4 +47,4 @@ Eigen supports cmake, so if sombody wants to integrate that support in the CMake
 Be sure Eigen is installed in your system. 
 Don't forget to add the path of the installed library when compiling, for example, if you want to compile example.cpp with g++:
 
-```g++ example.cpp -LCabernet/lib -lCabernet -I Cabernet/include```
+```g++ example.cpp -LCaberNet/lib -lCaberNet -I CaberNet/include```
