+++ include/CaberNet/layers.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "layers/layer.h"
+#include "layers/dense.h"
+#include "layers/relu.h"
