--- include/functions.h
@@ -1,16 +1,0 @@
-#ifndef FUNCTIONS_H
-#define FUNCTIONS_H
-
-#include "tensor.h"
-
-namespace net::function {
-
-Tensor linear(const Tensor& input, const Tensor& weight, const Tensor& bias);
-Tensor softmax(Tensor& input, int axis);
-Tensor log_softmax(Tensor&input, int axis);
-Tensor relu(const Tensor& input);
-
-} // namespace net
-
-
-#endif
