+++ src/functions.cpp
@@ -1,4 +1,5 @@
-#include "../include/functions.h"
+#include "../include/CaberNet/tensor.h"
+#include "../include/CaberNet/functions.h"
 #include "internals/internal_tensor.hpp"
 #include "internals/functions/internal_functions.hpp"
