+++ src/internals/functions/internal_function_logsoftmax.cpp
@@ -44,6 +44,7 @@ Tensor* LogSoftmax::forward() {
 
 void LogSoftmax::backward(Array* gradient) const {
     if (input()->requires_gradient()) {
+        throw std::runtime_error("Not implemented yet, if you want to contribute, you can start from here!");
         input()->backward(gradient);
     }
 }
